Refactor test app to use ES modules with single registration point
Replaces the previous pattern (global functions loaded via multiple
<script> tags in index.html) with ES modules and a single registration
file, mirroring purchases-flutter's `lib/test_cases.dart`.

Adding a new test case now only requires:
  1. Create www/js/screens/your_test_case.js exporting a `show*` function.
  2. Add an import + entry in www/js/test_cases.js.

No more <script> ordering gotchas, no implicit cross-file globals.

Made-with: Cursor
