+++ e2e-tests/MaestroTestApp/www/index.html
@@ -9,7 +9,6 @@
 <body>
     <div id="app"></div>
     <script src="cordova.js"></script>
-    <script src="js/screens/purchase_through_paywall.js"></script>
-    <script src="js/app.js"></script>
+    <script type="module" src="js/app.js"></script>
 </body>
 </html>
