+++ e2e-tests/MaestroTestApp/www/js/app.js
@@ -1,18 +1,4 @@
-// ---------------------------------------------------------------------------
-// Test case registry
-// To add a new test case:
-//   1. Create www/js/screens/your_test_case.js (see purchase_through_paywall.js)
-//   2. Add a <script> tag for it in index.html
-//   3. Add an entry to TEST_CASES below
-// ---------------------------------------------------------------------------
-
-var TEST_CASES = [
-    { title: 'Purchase through paywall', show: showPurchaseThroughPaywall }
-];
-
-// ---------------------------------------------------------------------------
-// Initialization
-// ---------------------------------------------------------------------------
+import { showTestCases } from './test_cases_screen.js';
 
 document.addEventListener('deviceready', function() {
     try {
@@ -33,33 +19,3 @@ document.addEventListener('deviceready', function() {
 
     showTestCases();
 }, false);
-
-// ---------------------------------------------------------------------------
-// Test Cases list screen
-// ---------------------------------------------------------------------------
-
-function showTestCases() {
-    var app = document.getElementById('app');
-    app.innerHTML = '<h1>Test Cases</h1>';
-    TEST_CASES.forEach(function(tc) {
-        var btn = document.createElement('button');
-        btn.textContent = tc.title;
-        btn.onclick = tc.show;
-        app.appendChild(btn);
-    });
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
-// ---------------------------------------------------------------------------
-
-function showError(msg) {
-    var el = document.getElementById('error-msg');
-    if (!el) {
-        el = document.createElement('p');
-        el.id = 'error-msg';
-        el.style.cssText = 'color:red;font-size:12px';
-        document.getElementById('app').appendChild(el);
-    }
-    el.textContent = msg;
-}
