+++ e2e-tests/MaestroTestApp/www/js/helpers.js
@@ -1,0 +1,10 @@
+export function showError(msg) {
+    var el = document.getElementById('error-msg');
+    if (!el) {
+        el = document.createElement('p');
+        el.id = 'error-msg';
+        el.style.cssText = 'color:red;font-size:12px';
+        document.getElementById('app').appendChild(el);
+    }
+    el.textContent = msg;
+}
